Working Jian Yang level classification stage.
At this point my network trained to recognise one gesture - "rotation clockwise" - and "something else". Kind of like the food classifier from the Silicon Valley: Recognises whether something is a hotdog or is not a hotdog. That said, the network is working and it only needs more thraining data on other gestures (and probably more neurons in the hidden layers: I probably need to use an ESP32 since I'm running out of ram on nano 33 ble). I'm also considering switching back to OpenLog if I change my microcontroller, since it is more compact that the naked SD card board.